# apps/ppaz-api-gateway/src/modules/secretarias/secretarias.controller.ts
import { Controller, Get, Post, Body, Patch, Put, Param, Delete } from '@nestjs/common';
import { HttpService } from '@nestjs/axios';
import { ConfigService } from '@nestjs/config';
import { ApiTags, ApiOperation, ApiResponse, ApiParam, ApiBody } from '@nestjs/swagger';
import { Observable } from 'rxjs';
import { map } from 'rxjs/operators';
import { CreateSecretariaDto } from './dto/create-secretaria.dto';
import { UpdateSecretariaDto } from './dto/update-secretaria.dto';

@ApiTags('secretarias')
@Controller('secretarias')
export class SecretariasController {
  private readonly authServiceUrl: string;

  constructor(
    private readonly httpService: HttpService,
    private readonly configService: ConfigService,
  ) {
    const host = this.configService.get<string>('PPP_AUTH_HOST');
    const port = this.configService.get<number>('PPP_AUTH_PORT');
    const isProduction = this.configService.get<string>('NODE_ENV') === 'production';
    
    this.authServiceUrl = isProduction 
      ? `https://${host}` 
      : `http://${host}:${port}`;
  }

  @Post()
  @ApiOperation({ summary: 'Crear una nueva secretaria', description: 'Registra una secretaria asociada a una escuela' })
  @ApiBody({ type: CreateSecretariaDto })
  @ApiResponse({ status: 201, description: 'Secretaria creada exitosamente' })
  @ApiResponse({ status: 400, description: 'Datos inválidos o escuela no existe' })
  create(@Body() createSecretariaDto: CreateSecretariaDto): Observable<any> {
    return this.httpService
      .post(`${this.authServiceUrl}/secretarias`, createSecretariaDto)
      .pipe(map((response) => response.data));
  }

  @Get()
  @ApiOperation({ summary: 'Listar todas las secretarias', description: 'Obtiene todas las secretarias con información de su escuela y facultad' })
  @ApiResponse({ status: 200, description: 'Lista de secretarias obtenida exitosamente' })
  findAll(): Observable<any> {
    return this.httpService
      .get(`${this.authServiceUrl}/secretarias`)
      .pipe(map((response) => response.data));
  }

  @Get(':id')
  @ApiOperation({ summary: 'Obtener secretaria por ID', description: 'Busca una secretaria específica por su UUID' })
  @ApiParam({ name: 'id', description: 'UUID de la secretaria', type: 'string' })
  @ApiResponse({ status: 200, description: 'Secretaria encontrada' })
  @ApiResponse({ status: 404, description: 'Secretaria no encontrada' })
  findOne(@Param('id') id: string): Observable<any> {
    return this.httpService
      .get(`${this.authServiceUrl}/secretarias/${id}`)
      .pipe(map((response) => response.data));
  }

  @Get('escuela/:idEscuela')
  @ApiOperation({ summary: 'Listar secretarias por escuela', description: 'Obtiene todas las secretarias de una escuela específica' })
  @ApiParam({ name: 'idEscuela', description: 'UUID de la escuela', type: 'string' })
  @ApiResponse({ status: 200, description: 'Lista de secretarias obtenida' })
  findByEscuela(@Param('idEscuela') idEscuela: string): Observable<any> {
    return this.httpService
      .get(`${this.authServiceUrl}/secretarias/escuela/${idEscuela}`)
      .pipe(map((response) => response.data));
  }

  @Put(':id')
  @Patch(':id')
  @ApiOperation({ summary: 'Actualizar secretaria', description: 'Actualiza los datos de una secretaria existente' })
  @ApiParam({ name: 'id', description: 'UUID de la secretaria', type: 'string' })
  @ApiBody({ type: UpdateSecretariaDto })
  @ApiResponse({ status: 200, description: 'Secretaria actualizada exitosamente' })
  @ApiResponse({ status: 404, description: 'Secretaria no encontrada' })
  update(@Param('id') id: string, @Body() updateSecretariaDto: UpdateSecretariaDto): Observable<any> {
    return this.httpService
      .patch(`${this.authServiceUrl}/secretarias/${id}`, updateSecretariaDto)
      .pipe(map((response) => response.data));
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Eliminar secretaria', description: 'Elimina una secretaria del sistema' })
  @ApiParam({ name: 'id', description: 'UUID de la secretaria', type: 'string' })
  @ApiResponse({ status: 200, description: 'Secretaria eliminada exitosamente' })
  @ApiResponse({ status: 404, description: 'Secretaria no encontrada' })
  remove(@Param('id') id: string): Observable<any> {
    return this.httpService
      .delete(`${this.authServiceUrl}/secretarias/${id}`)
      .pipe(map((response) => response.data));
  }
}