# swagger-auth.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/secretarias": {
            "post": {
                "operationId": "SecretariasController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSecretariaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Secretaria creada exitosamente"
                    }
                },
                "summary": "Crear una nueva secretaria",
                "tags": [
                    "Secretarias"
                ]
            },
            "get": {
                "operationId": "SecretariasController_findAll",
                "parameters": [
                    {
                        "name": "escuela",
                        "required": false,
                        "in": "query",
                        "description": "Filtrar por ID de escuela",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de secretarias"
                    }
                },
                "summary": "Obtener todas las secretarias",
                "tags": [
                    "Secretarias"
                ]
            }
        },
        "/secretarias/{id}": {
            "get": {
                "operationId": "SecretariasController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secretaria encontrada"
                    }
                },
                "summary": "Obtener una secretaria por ID",
                "tags": [
                    "Secretarias"
                ]
            },
            "patch": {
                "operationId": "SecretariasController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSecretariaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Secretaria actualizada"
                    }
                },
                "summary": "Actualizar una secretaria",
                "tags": [
                    "Secretarias"
                ]
            },
            "delete": {
                "operationId": "SecretariasController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Secretaria eliminada"
                    }
                },
                "summary": "Eliminar una secretaria",
                "tags": [
                    "Secretarias"
                ]
            }
        },
...
    },
    "info": {
        "title": "PPP Auth Service API",
        "description": "Autenticación y Autorización - Usuarios, Roles, Privilegios",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "CreateSecretariaDto": {
                "type": "object",
                "properties": {
                    "usuarioId": {
                        "type": "string",
                        "description": "ID del usuario asociado",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "idEscuela": {
                        "type": "string",
                        "description": "ID de la escuela profesional",
                        "example": "123e4567-e89b-12d3-a456-426614174001"
                    }
                },
                "required": [
                    "usuarioId",
                    "idEscuela"
                ]
            },
            "UpdateSecretariaDto": {
                "type": "object",
                "properties": {
                    "idEscuela": {
                        "type": "string",
                        "description": "ID de la escuela profesional",
                        "example": "123e4567-e89b-12d3-a456-426614174001"
                    }
                }
            },
...
        }
    }
}